reorg toolbar notification
